HLRC: split graph request converters
In an effort to encapsulate the different clients, the request
converters are being shuffled around. This splits the GraphClient
request converters.
